# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: QuakeAlert REST API
  version: 1.0.0
  description: >
    Kontrak REST resmi QuakeAlert (contract-first, ADR-0004). Base URL selalu HTTPS.
    Satuan kanonik: PGA=gal (cm/s^2), timestamp=ms epoch UTC (int64), jarak=km, RSSI=dBm.
    Semua endpoint memerlukan Bearer JWT (anonymous) KECUALI:
    `GET /healthz` (probe publik), `POST /api/v1/auth/anonymous`
    (publik — bootstrap identitas) dan `GET /api/v1/events`
    (auth opsional; bila token dikirim ia wajib valid).
    Empat endpoint berikut adalah permukaan integrasi resmi klien Android
    (alur first-launch → lokasi → push → riwayat); panduan integrasi ada di
    docs/CLIENT_SPEC.md: `POST /api/v1/auth/anonymous`,
    `PUT /api/v1/users/location`, `PUT /api/v1/users/fcm-token`,
    `GET /api/v1/events`. Endpoint lainnya dipakai firmware (provisioning) dan
    tooling web (sensors, reroll pseudonym).
servers:
  - url: https://api.quakealert.id
    description: Produksi
  - url: http://localhost:8080
    description: Dev lokal (docker-compose)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/nodes/{stationID}/admin-designate:
    post:
      tags: [admin]
      summary: Menunjuk satu node terverifikasi sebagai Admin Node (D-036, PROPOSED)
      description: >
        Aksi eksplisit operator; tidak ada penunjukan otomatis oleh migrasi
        maupun server. Syarat: station_id berpola benar, baris dikenal,
        verified = true, dan belum ada node lain yang memegang kapabilitas
        (maksimum satu pada deployment saat ini). Idempoten: menunjuk ulang
        pemegang saat ini menjawab 200 tanpa efek. station_id yang salah bentuk
        atau tidak dikenal menjawab 404 dan tidak pernah menyentuh basis data.
      operationId: designateAdminNode
      security:
        - adminKey: []
      parameters:
        - name: stationID
          in: path
          required: true
          schema:
            type: string
            pattern: '^NODE-[0-9A-F]{8}$'
      responses:
        '200':
          description: Node ditunjuk sebagai Admin Node
          content:
            application/json:
              schema:
                type: object
                properties:
                  station_id:
                    type: string
                    example: NODE-163A149F
                  is_admin_node:
                    type: boolean
                    example: true
        '401':
          description: Header X-Admin-Key tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: station_id salah bentuk atau tidak dikenal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NODE_NOT_FOUND
                message: station_id tidak ditemukan
        '409':
          description: Node belum terverifikasi, atau sudah ada Admin Node lain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ADMIN_NODE_EXISTS
                message: sudah ada Admin Node aktif
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/admin/nodes/{stationID}/admin-revoke:
    post:
      tags: [admin]
      summary: Mencabut kapabilitas Admin Node satu node (D-036, PROPOSED)
      description: >
        Idempoten: mencabut node yang memang bukan admin tetap menjawab 200
        dengan is_admin_node=false. station_id yang salah bentuk atau tidak
        dikenal menjawab 404 dan tidak pernah menyentuh basis data.
      operationId: revokeAdminNode
      security:
        - adminKey: []
      parameters:
        - name: stationID
          in: path
          required: true
          schema:
            type: string
            pattern: '^NODE-[0-9A-F]{8}$'
      responses:
        '200':
          description: Kapabilitas Admin Node dicabut
          content:
            application/json:
              schema:
                type: object
                properties:
                  station_id:
                    type: string
                    example: NODE-163A149F
                  is_admin_node:
                    type: boolean
                    example: false
        '401':
          description: Header X-Admin-Key tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: station_id salah bentuk atau tidak dikenal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NODE_NOT_FOUND
                message: station_id tidak ditemukan
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Shared secret operator, minimal 32 byte, dibandingkan dengan
        crypto/subtle.ConstantTimeCompare. Dipakai HANYA oleh /api/v1/admin/*.
  responses:
...
    InternalError:
      description: Kesalahan internal server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            Kode mesin yang stabil. Nilai yang dipakai server:
            INVALID_ARGUMENT, UNAUTHENTICATED, STATION_ALREADY_EXISTS,
            RATE_LIMITED, INTERNAL, dan — khusus /api/v1/nodes/revoke —
            SECRET_MISMATCH, NODE_NOT_FOUND, NODE_VERIFIED — serta, khusus
            endpoint admin tracker saat tracker nonaktif, TRACKER_DISABLED;
            khusus chat, PERMISSION_DENIED; khusus drill saat dependensi
            belum siap, UNAVAILABLE; khusus designate/revoke Admin Node
            (D-036, PROPOSED), NODE_NOT_VERIFIED, ADMIN_NODE_EXISTS.
          enum:
            - INVALID_ARGUMENT
            - UNAUTHENTICATED
            - STATION_ALREADY_EXISTS
            - RATE_LIMITED
            - INTERNAL
            - SECRET_MISMATCH
            - NODE_NOT_FOUND
            - NODE_VERIFIED
            - TRACKER_DISABLED
            - PERMISSION_DENIED
            - UNAVAILABLE
            - NODE_NOT_VERIFIED
            - ADMIN_NODE_EXISTS
          example: INVALID_ARGUMENT
        message:
          type: string
          example: latitude di luar rentang -90..90
...